refactor(cbl): move hook helper code into src/install
Move download.dart from hook/src to packages/cbl/src/install to keep
the hook directory clean. Update build.dart import path accordingly.
